feat: modificado properties para criar o banco de dados no H2

+++ clinica-odontologica/src/main/java/com/dh/clinicaodontologica/controller/EnderecoController.java
@@ -1,6 +1,6 @@
 package com.dh.clinicaodontologica.controller;
 
-import com.dh.clinicaodontologica.dto.EnderecoDTO;
+import dto.EnderecoDTO;
 import com.dh.clinicaodontologica.service.impl.EnderecoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -25,7 +25,7 @@ public class EnderecoController {
     }
 
     @PutMapping("/update/{id}")
-    public EnderecoDTO update(@RequestBody EnderecoDTO enderecoDTO, @PathVariable int id) {
+    public EnderecoDTO update(@RequestBody EnderecoDTO enderecoDTO, @PathVariable Long id) {
         return enderecoService.update(enderecoDTO, id);
     }
 }
